+++ CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - DualSense / DualShock 4: Report Acceleromenter and Gyro data
 - Steam Controller: Add support for Steam Controller.
   - Disables "lizard" mode
-  - Dpad, buttons, triggers, thumbstick supported.
-  - Gyro/Accel/Right pad: not supported ATM
+  - Dpad, buttons, triggers, thumbstick, right pad supported.
+  - Gyro/Accel: not supported ATM
 
 ### Fixed
 - Xbox Adaptive Controller: Works as expeted. Removed "unsupported usage" messages.