+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - ???
+## [v3.7.1] - 2023-04-30
 ### New
 - DualSense / DualShock 4: Report Acceleromenter and Gyro data
 - Steam Controller: Add support for Steam Controller.
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Xbox Adaptive Controller: Works as expeted. Removed "unsupported usage" messages.
+- Arduino / NINA / CircuitPython: Add APIs to read gyro / accel.
 
 ## [v3.7.0] - 2023-04-17
 ### New